Refactor module structure from Ruby::Pericope to Pericope
- Changed primary module name from Ruby::Pericope to just Pericope
- Moved files from lib/ruby/pericope/ to lib/pericope/
- Updated all module declarations to remove Ruby namespace wrapper
- Updated all require statements to reflect new directory structure
- Updated gemspec to reference correct module and version path
- Updated README.md and example files to use new module structure
- Updated all test files to use new module structure
- Removed old lib/ruby and spec/ruby directories
- Fixed code style issues with RuboCop auto-correct
- All tests passing (172 examples, 0 failures)
