+++ lib/pericope.rb
@@ -7,10 +7,8 @@ require_relative "pericope/book"
 require_relative "pericope/verse_ref"
 require_relative "pericope/pericope"
 
-module Ruby
-  module Pericope
-    class Error < StandardError; end
+module Pericope
+  class Error < StandardError; end
 
-    # Your code goes here...
-  end
+  # Your code goes here...
 end
