--- lib/ruby/pericope/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Ruby
-  module Pericope
-    VERSION = "0.1.0"
-  end
-end
